Adding character and word screens

+++ src/MorphShared/Models/WordFrequency.cs
@@ -3,6 +3,6 @@
     public class WordFrequency
     {
         public string Word { get; set; }
-        public int Frequency { get; set; }
+        public double Frequency { get; set; }
     }
 }
